Fix before block on reformat_test
It was working locally because the spec/tmp_files directory
had been manually created.

(We aren't removing it in an after block in case we want to examine
the file after a failing test.)

+++ spec/reformat_spec.rb
@@ -5,7 +5,7 @@ require_relative '../lib/formatter'
 RSpec.describe "cli for formatter" do
 
   before do
-    FileUtils.rm_r("spec/tmp_files")
+    FileUtils.rm_r("spec/tmp_files") if Dir.exist?("spec/tmp_files")
     FileUtils.mkdir("spec/tmp_files")
   end
